// apps/admin/src/generated/api/schemas/departmentTreeNodeDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentTreeNodeDtoStatus } from "./departmentTreeNodeDtoStatus";

export interface DepartmentTreeNodeDto {
  id: string;
  code: string;
  name: string;
  /** @nullable */
  parentId?: string | null;
  status: DepartmentTreeNodeDtoStatus;
  sortOrder: number;
  children: DepartmentTreeNodeDto[];
}