// apps/admin/src/generated/api/schemas/updateDepartmentDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDepartmentDtoStatus } from "./updateDepartmentDtoStatus";

export interface UpdateDepartmentDto {
  /** @maxLength 80 */
  code?: string;
  /** @maxLength 120 */
  name?: string;
  /** @nullable */
  parentId?: string | null;
  status?: UpdateDepartmentDtoStatus;
  /** @minimum 0 */
  sortOrder?: number;
  /**
   * @maxLength 500
   * @nullable
   */
  description?: string | null;
}